refactor: seperate update logic
